Fixed inventory duplication bug.
The duplication occurred when keepInventory=true and a player died
with a custom item on the inventory.

On beta (0.0.x) versions this would only happen with the King's Crown.

On versions prior to this one and after the major API refactoring
this would happen with all custom items.

Thanks to Sopas, an admin at minept.net.

The bug is now fixed :)